+++ tests/SessionAuthorizationTest.fs
@@ -52,6 +52,13 @@ let provideAuthorizations: AuthorizationTestCase<string, string, string> list = 
         | "data" -> AsyncResult.ofSuccess "response"
         | _ -> AsyncResult.ofSuccess "wrong-response"
 
+    let enforceScope =
+        let model = AuthorizationTest.model "rbac_model.conf"
+        let policy = AuthorizationTest.policy "adminConsole.csv"
+        let enforcer = Authorization.createEnforcer model policy |> okOrFail
+
+        EnforceScope.prepare enforcer (_.Username >> Option.defaultValue "anonymous" >> Subject)
+
     {
         Description = "should authorize action with login"
         Authorization = authorization
@@ -108,6 +115,37 @@ let provideAuthorizations: AuthorizationTestCase<string, string, string> list = 
         Description = "should NOT authorize action with group and not pass a username"
         Authorization = authorization
         Authorize = Authorize.withGroup (PermissionGroup "admin")
+        Action =
+            Authorize.Action.RequestWithUsername (fun username ->
+                failtestf "Username should not be passed in."
+            )
+        Request = {
+            Token = SecurityToken jwt
+            RequestData = "data"
+        }
+        ValidateToken = ignore  // todo - maybe later
+        Expected = Error (SecuredRequestError.AuthorizationError "Action is not granted! You are not authorized for this action.")
+    }
+    {
+        Description = "should authorize action with scope"
+        Authorization = authorization
+        Authorize = Authorize.withScope enforceScope (AuthorizationTest.scope "data:read")
+        Action =
+            Authorize.Action.RequestWithUsername (fun username ->
+                Expect.equal (Username "user") username "Username should be passed in."
+                action
+            )
+        Request = {
+            Token = SecurityToken jwt
+            RequestData = "data"
+        }
+        ValidateToken = ignore  // todo - maybe later
+        Expected = Ok "response"
+    }
+    {
+        Description = "should NOT authorize action with scope"
+        Authorization = authorization
+        Authorize = Authorize.withScope enforceScope (AuthorizationTest.scope "data:write")
         Action =
             Authorize.Action.RequestWithUsername (fun username ->
                 failtestf "Username should not be passed in."